+++ packages/components/src/Menu/MenuList.tsx
@@ -101,9 +101,7 @@ export const MenuListInternal = forwardRef(
     { children, compact, disabled, pin, placement, ...props }: MenuListProps,
     forwardedRef: Ref<HTMLUListElement>
   ) => {
-    const { groupedMenusRef, id, isOpen, setOpen, triggerElement } = useContext(
-      MenuContext
-    )
+    const { id, isOpen, setOpen, triggerElement } = useContext(MenuContext)
 
     const [renderIconPlaceholder, setRenderIconPlaceholder] = useState(false)
 
@@ -148,7 +146,6 @@ export const MenuListInternal = forwardRef(
     const isMenu = isOpen !== undefined
     const { popover } = usePopover({
       content: menuList,
-      groupedPopoversRef: groupedMenusRef,
       isOpen,
       pin,
       placement,